Refactor partitioning class
